[DsComparisonPage|DsComparisonGrid] modularized grid state handler to a different component
