+++ metaspace/webapp/src/modules/Datasets/comparison/DatasetComparisonPage.scss
@@ -15,6 +15,7 @@
 
   .ds-collapse{
     margin: 0 8px;
+    min-height: 100px;
     width: calc(100% - 16px);
     .el-collapse-item__header{
       background-color: #F1F5F8;
